+++ src/main/java/Entities/InterestTag.java
@@ -1,4 +1,4 @@
-package Users;
+package Entities;
 //TODO: to be implemented in stage 2. Need tagDataBase.
 //Tags for of interests. We will decide on a list of interest tags for users to choose from. Students will also be able
 // to see others who have the same tag selected.
