+++ src/main/java/Entities/Label.java
@@ -1,4 +1,4 @@
-package Users;
+package Entities;
 
 // Labels of status/moods/intents. We will create a list for students to choose from, and display them alongside matched results.
 //e.g. Want to collaborate, Want to discuss, Want to review together, etc.
