+++ src/main/java/GUI/HomeFrame.java
@@ -1,5 +1,7 @@
 package GUI;
 
+import UIController.UIController;
+
 import javax.swing.*;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
