+++ src/main/java/GUI/RegisterFrame.java
@@ -1,5 +1,7 @@
 package GUI;
 
+import UIController.UIController;
+
 import javax.swing.*;
 import java.awt.*;
 import java.awt.event.ActionEvent;
