+++ src/main/java/GUI/TagMatchFrame.java
@@ -1,17 +1,12 @@
 package GUI;
-import Users.Student;
-import com.sun.tools.jconsole.JConsoleContext;
-import useCases.TagMatchManager;
+
+import UIController.UIController;
 
 import javax.swing.*;
-import javax.swing.event.ListSelectionEvent;
-import javax.swing.event.ListSelectionListener;
-import java.awt.*;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.awt.event.ItemEvent;
 import java.awt.event.ItemListener;
-import java.io.IOException;
 
 public class TagMatchFrame extends JFrame implements ActionListener, ItemListener {
