+++ src/main/java/GUI/TagSelectFrame.java
@@ -1,17 +1,10 @@
 package GUI;
-import Users.Student;
-import com.sun.tools.jconsole.JConsoleContext;
-import useCases.TagMatchManager;
+
+import UIController.UIController;
 
 import javax.swing.*;
-import javax.swing.event.ListSelectionEvent;
-import javax.swing.event.ListSelectionListener;
-import java.awt.*;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
-import java.awt.event.ItemEvent;
-import java.awt.event.ItemListener;
-import java.io.IOException;
 import java.util.ArrayList;
 
 public class TagSelectFrame extends JFrame implements ActionListener{
