--- src/main/java/testFolder/test222.java
@@ -1,4 +1,0 @@
-package testFolder;
-
-public class test222 {
-}
